creates EditorAndOutput component
