+++ src/components/Main.js
@@ -1,25 +1,14 @@
 import React from "react";
-import SplitPane from "react-split-pane";
-import OutputContainer from "./Output/OutputContainer.js";
-import TextEditorContainer from "./TextEditor/containers/TextEditorContainer.js";
-import DropdownButtonContainer from "./common/containers/DropdownButtonContainer";
+
 import * as fetch from "../lib/fetch.js";
 import * as cookies from "../lib/cookies.js";
-import SketchesPageContainer from "./Sketches/containers/SketchesContainer";
-import "styles/Main.scss";
+
 import ProfilePanelContainer from "./common/containers/ProfilePanelContainer";
+import EditorAndOutput from "./EditorAndOutput/EditorAndOutput";
 
-import { Redirect } from "react-router-dom";
-import { EDITOR_WIDTH_BREAKPOINT, CODE_AND_OUTPUT, CODE_ONLY, OUTPUT_ONLY } from "../constants";
-import CodeDownloader from "../util/languages/CodeDownloader";
+import { EDITOR_WIDTH_BREAKPOINT, CODE_AND_OUTPUT, CODE_ONLY } from "../constants";
 
-import { PANEL_SIZE } from "../constants";
-import "codemirror/lib/codemirror.css";
-import "codemirror/theme/material.css";
-import "codemirror/theme/duotone-light.css";
-import "styles/CustomCM.scss";
-import "styles/Resizer.scss";
-import "styles/Editor.scss";
+import "styles/Main.scss";
 
 /**------Props-------
  * togglePanel: function to call when you want the Profile Panel to disappear/reapper
@@ -76,7 +65,7 @@ class Main extends React.Component {
     });
   };
 
-  handleSave = event => {
+  handleSave = () => {
     if (!this.props.dirty) return; // Don't save if not dirty (unedited)
     this.setState({
       saveText: "Saving...",
@@ -98,14 +87,6 @@ class Main extends React.Component {
     this.props.cleanCode(this.props.mostRecentProgram); // Set code's "dirty" state to false
   };
 
-  handleDownload = () => {
-    CodeDownloader.download(this.props.name, this.props.language, this.props.code);
-  };
-
-  renderDropdown = () => <DropdownButtonContainer />;
-
-  renderSketchesPage = () => <SketchesPageContainer />;
-
   renderContent = () => {
     switch (this.props.contentType) {
       case "sketches":
@@ -116,85 +97,36 @@ class Main extends React.Component {
     }
   };
 
-  renderCodeAndOutput = () => (
-    <SplitPane
-      resizerStyle={{
-        height: "67px",
-        borderLeft: "2px solid #333",
-        borderRight: "2px solid #333",
-        width: "10px",
-      }}
-      pane1Style={this.state.pane1Style}
-      //functions called when you start and finish a drag
-      //removes and re-addsthe transition effect on the first panel when manually resizing
-      onDragStarted={() => this.setState({ pane1Style: {} })}
-      onDragFinished={() => this.setState({ pane1Style: { transition: "width .5s ease" } })}
-      split="vertical" //the resizer is a vertical line (horizontal means resizer is a horizontal bar)
-      minSize={
-        (this.props.panelOpen ? this.props.screenWidth - PANEL_SIZE : this.props.screenWidth) * 0.33
-      } //minimum size of code is 33% of the remaining screen size
-      maxSize={
-        (this.props.panelOpen ? this.props.screenWidth - PANEL_SIZE : this.props.screenWidth) * 0.75
-      } //max size of code is 75% of the remaining screen size
-      size={
-        ((this.props.panelOpen ? this.props.screenWidth - PANEL_SIZE : this.props.screenWidth) /
-          5) *
-        3
-      } //the initial size of the text editor section
-      allowResize={true}
-    >
-      {this.renderCode()}
-      {this.renderOutput()}
-    </SplitPane>
-  );
-
-  updateViewMode = viewMode => {
-    this.setState({ viewMode });
-  };
-
-  renderCode = () => (
-    <TextEditorContainer
-      key={this.props.mostRecentProgram}
+  renderEditor = () => (
+    <EditorAndOutput
+      // view mode
       viewMode={this.state.viewMode}
-      updateViewMode={this.updateViewMode}
-      handleSave={this.handleSave}
-      saveText={this.state.saveText}
-      screenHeight={this.props.screenHeight}
-      screenWidth={this.props.screenWidth}
-      handleDownload={this.handleDownload}
+      updateViewMode={viewMode => this.setState({ viewMode })}
+      // theme
       theme={this.state.theme}
+      // sizing
+      left={this.props.left}
+      screenWidth={this.props.screenWidth}
+      screenHeight={this.props.screenHeight}
+      // view only trigger
+      viewOnly={false}
+      // pane
+      panelOpen={this.props.panelOpen}
+      pane1Style={this.state.pane1Style}
+      changePane1Style={newStyle => this.setState(newStyle)}
+      // program information
+      mostRecentProgram={this.props.mostRecentProgram}
+      language={this.state.language}
+      code={this.state.code}
+      programid={this.props.programid}
+      sketchName={this.state.sketchName}
+      // save handler
+      saveText={this.state.saveText}
+      handleSave={this.handleSave}
     />
   );
-
-  renderOutput = () => (
-    <OutputContainer
-      viewMode={this.state.viewMode}
-      updateViewMode={this.updateViewMode}
-      isSmall={this.props.screenWidth <= EDITOR_WIDTH_BREAKPOINT}
-    />
-  );
-
-  renderEditor = () => {
-    const codeStyle = {
-      width: this.props.screenWidth - (this.props.left || 0),
-      height: this.props.screenHeight,
-    };
-
-    switch (this.state.viewMode) {
-      case CODE_ONLY:
-        return <div style={codeStyle}>{this.renderCode()}</div>;
-      case OUTPUT_ONLY:
-        return <div style={codeStyle}>{this.renderOutput()}</div>;
-      default:
-        return this.renderCodeAndOutput();
-    }
-  };
 
   render() {
-    if (this.state.redirect) {
-      return <Redirect to={this.state.redirect} />;
-    }
-
     const codeStyle = {
       left: this.props.left || 0,
       width: this.props.screenWidth - (this.props.left || 0),
